genl: Always call subdissector
Commit 61c5e8e76d21 ("genl: make subdissectors responsible for header")
changed the generic netlink dissector to only call a sub-dissector if
there is a payload after the generic netlink header.

However, there are commands in certain generic netlink families (e.g.,
"NET_DM") that do not have any payload. This means that the command will
not be dissected by the subdissector, as it will never be invoked.

Change the generic netlink dissector to always call a subdissector, if
it is present.

Change-Id: I4d2f48531dee92b11dc45000081a8d2d3dd875c6
Signed-off-by: Ido Schimmel <idosch@mellanox.com>